--- src/icons/CircleHalf.tsx
@@ -1,14 +1,0 @@
-interface Props {
-	width?: number;
-	height?: number;
-	className?: string;
-}
-
-export function CircleHalf({ width, height, className }: Props) {
-	// https://icons.getbootstrap.com/icons/circle-half/
-	return (
-		<svg xmlns="http://www.w3.org/2000/svg" className={className} width={width} height={height} fill="currentColor" viewBox="0 0 16 16">
-			<path d="M8 15A7 7 0 1 0 8 1zm0 1A8 8 0 1 1 8 0a8 8 0 0 1 0 16" />
-		</svg>
-	);
-}
